KT-SFT/requirements-sft.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
attrs==25.3.0
colorama==0.4.6
conda-pack==0.8.1
datasets==3.6.0
dill==0.3.8
einops==0.8.1
frozenlist==1.6.0
graphviz==0.20.3
joblib==1.5.1
multidict==6.4.4
multiprocess==0.70.16
nltk==3.9.1
nvidia-cufile-cu12==1.11.1.6
pandas==2.2.3
peft==0.14.0
propcache==0.3.1
pyarrow==20.0.0
python-dateutil==2.9.0.post0
python_helper==0.3.74
pytz==2025.2
rouge_score==0.1.2
six==1.17.0
tabulate==0.9.0
thop==0.1.1.post2209072238
torchviz==0.0.3
tzdata==2025.2
xxhash==3.5.0
yarl==1.20.0
torchviz